# Source/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AFCU Bank API Library",
    "version": "1.0.0",
    "description": "API Documentation for Armed Forces Credit Union (AFCU) Bank project"
  },
  "host": "54.92.189.15:3000",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Firebase JWT token as 'Bearer [your token]'"
    }
  },
  "paths": {
    "/account/create/{name}/{email}/{password}/{balance}": {
      "get": {
        "tags": ["Account"],
        "summary": "Create a new user account",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "name", "in": "path", "required": true, "type": "string" },
          { "name": "email", "in": "path", "required": true, "type": "string" },
          { "name": "password", "in": "path", "required": true, "type": "string" },
          { "name": "balance", "in": "path", "required": true, "type": "number" }
        ],
        "responses": {
          "200": {
            "description": "Account successfully created",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/account/login/{email}/{password}": {
      "get": {
        "tags": ["Account"],
        "summary": "Log in to an account",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "email", "in": "path", "required": true, "type": "string" },
          { "name": "password", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "Logged in successfully",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/account/logout": {
      "get": {
        "tags": ["Account"],
        "summary": "Logout of an account",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/account/all": {
      "get": {
        "tags": ["Account"],
        "summary": "Get all accounts",
        "security": [{ "Bearer": [] }],
        "responses": {
          "200": {
            "description": "Accounts fetched successfully",
            "schema": {
              "type": "array",
              "items": { "type": "object" }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/account/deposit/{email}/{amount}": {
      "get": {
        "tags": ["Account"],
        "summary": "Deposit to an account",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "email", "in": "path", "required": true, "type": "string" },
          { "name": "amount", "in": "path", "required": true, "type": "number" }
        ],
        "responses": {
          "200": {
            "description": "Deposit successful",
            "schema": { "type": "object" }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/account/withdraw/{email}/{amount}": {
      "get": {
        "tags": ["Account"],
        "summary": "Withdraw from an account",
        "security": [{ "Bearer": [] }],
        "parameters": [
          { "name": "email", "in": "path", "required": true, "type": "string" },
          { "name": "amount", "in": "path", "required": true, "type": "number" }
        ],
        "responses": {
          "200": {
            "description": "Withdrawal successful",
            "schema": { "type": "object" }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/account/balance/{email}": {
      "get": {
        "tags": ["Account"],
        "summary": "Get balance of an account",
        "security": [{ "Bearer": [] }],
        "parameters": [{ "name": "email", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": {
            "description": "Balance fetched successfully",
            "schema": { "type": "object" }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}